Removed some of the Board class's unneeded global variables

+++ Swarm-Tool/src/strategies/AbstractStrategy.java
@@ -8,11 +8,11 @@ import other.SwarmAgent;
 
 //Goal strategies are children of AbstractStrategy
 public abstract class AbstractStrategy {
-	public abstract CellDisplayPolarity Layer2 (CellDisplayBase[][] layer1, int cellSize, int row, int col, Cell[] neighbor);
+	public abstract CellDisplayPolarity Layer2 (CellDisplayBase[][] layer1, double cellSize, int row, int col, Cell[] neighbor);
 	//MODIFICATION
 	//new parameter, agent, is added to the method
-	public abstract CellDisplayPolarity Layer2 (SwarmAgent agent, CellDisplayBase[][] layer1, int cellSize, int row, int col, Cell[] neighbor);
-	public abstract void logic (SwarmAgent agent, CellDisplayBase[][] layer1, CellDisplayPolarity[][] layer2, Cell[] neighbors, int cellSize);
+	public abstract CellDisplayPolarity Layer2 (SwarmAgent agent, CellDisplayBase[][] layer1, double cellSize, int row, int col, Cell[] neighbor);
+	public abstract void logic (SwarmAgent agent, CellDisplayBase[][] layer1, CellDisplayPolarity[][] layer2, Cell[] neighbors, double cellSize);
 	
 	//MODIFICATION #7:
 	//By Morgan Might
